monobj: checkCol single-precision debugRadius multiply (93.16->93.18%)

+++ src/monobj.cpp
@@ -2277,8 +2277,7 @@ void CGMonObj::checkCol(int flags, float rotY, float distance, float* hitScale, 
 		}
 
 		if (static_cast<double>(kMonObjZero) != static_cast<double>(halfAngle)) {
-			float debugRadius = static_cast<float>(
-				static_cast<double>(coneLength) * static_cast<double>(static_cast<float>(tan(static_cast<double>(halfAngle)))));
+			float debugRadius = coneLength * static_cast<float>(tan(static_cast<double>(halfAngle)));
 			CFlat.AddDebugDrawCC(&coneStart, reinterpret_cast<Vec*>(&move), debugRadius, 0, didHit);
 		}
 	}
